update(devide files): iterator.hpp -> iterator_traits.hpp & reverse_iterator.hpp
